asdfghj

+++ lib/features/screen/medicines.dart
@@ -73,8 +73,8 @@ class _MedicinePageState extends ConsumerState<MedicinePage> {
             image: medImage.toString(),
             name: nameController.text,
              ml: mlController.text,
-            rate: int.parse(rateController.text),
-            off: int.parse(offController.text,),
+            rate: double.parse(rateController.text),
+            off: double.parse(offController.text,),
             id:  idController.text));
   }
   @override
